partial update
