+++ smcnuts/smc_sampler.py
@@ -2,9 +2,10 @@ from time import time
 
 import autograd.numpy as np
 from tqdm import tqdm
-from scipy.special import logsumexp
+
 from smcnuts.tempering.adaptive_tempering import AdaptiveTempering
 from .utils.CheckAttributes import *
+from samples import Samples
 
 class SMCSampler():
     """Hamiltonian Monte Carlo SMC Sampler
@@ -40,7 +41,9 @@ class SMCSampler():
         self.sample_proposal = sample_proposal  # Initial sample proposal distribution
         self.tempering = tempering  # Tempering scheme
         self.lkernel = lkernel  # L-kernel distribution
- 
+        self.samples = Samples(self.N, self.sample_proposal)
+
+
         self.verbose = verbose  # Show stdout
         self.rng = rng  # Random number generator        
 
@@ -56,70 +59,7 @@ class SMCSampler():
         self.acceptance_rate = np.zeros(self.K)
         self.run_time = None
 
-    def normalise_weights(self, logw):
-        """
-        Normalises the sample weights
 
-        Args:
-            logw: A list of sample weights on the log scale
-
-        Returns:
-            A list of normalised weights
-
-        """
-
-        index = ~np.isneginf(logw)
-
-        log_likelihood = logsumexp(logw[index])
-
-        # Normalise the weights
-        wn = np.zeros_like(logw)
-        wn[index] = np.exp(logw[index] - log_likelihood)
-
-        return wn, log_likelihood  # type: ignore
-
-    def calculate_ess(self, wn):
-        """
-        Calculate the effective sample size using the normalised
-        sample weights.
-
-        Args:
-            wn: A list of normalised sample weights
-
-        Return:
-            The effective sample size
-        """
-
-        ess = 1 / np.sum(np.square(wn))
-
-        return ess
-
-    def resample(self, x, wn, log_likelihood):
-        """
-        Resamples samples and their weights from the specified indexes. If running the SMC sampler
-        in parallel using MPI, we resample locally on rank zero and then scatter the resampled
-        samples to the other ranks.
-
-        Args:
-            x: A list of samples to resample
-            wn: A list of normalise sample weights to resample
-            indexes: A list of the indexes of samples and weights to resample
-
-        Returns:
-            x_new: A list of resampled samples
-            logw_new: A list of resampled weights
-        """
-
-        # Resample x
-        i = np.linspace(0, self.N-1, self.N, dtype=int)
-        i_new = self.rng.choice(i, self.N, p=wn)
-        x_new = x[i_new]
-
-        # Determine new weights
-        # logw_new = np.log(np.ones(self.N_local)) - self.N_local
-        logw_new = (np.ones(self.N) * log_likelihood) - np.log(self.N)
-
-        return x_new, logw_new
 
     def estimate(self, x, wn):
         """
